JAMES-2214 Introduce helper class for saving and sending emails with JMAP
